Projects/chat-app/fast-backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
passlib[bcrypt]
python-jose[cryptography]
